integration tests: Move example_state() to examplelib
